+++ CHANGES.rst
@@ -6,6 +6,8 @@ Development
 
 - Started some documentation.
 - Set default root factory to ``royal.resource.Root``
+- Exception view which catches all exception is only added when ``debug=false``
+  in application settings.
 
 0.7.1
 -----